refactor: banner 컴포넌트에서 커스텀 훅 useMixedContent로 api 로직 분리
